Fuzzy game filter, no match error, and debug info
Pickysaurus mentioned that users might make typos in the game name, so now the search is fuzzy, allowing up to an approximate 1/3 difference between the input and the actual game name.

The game is now also included alongside the original query wherever it appears, and a game string that does not match any game registered with Nexus will inform the user of such.